pyproject.toml
[project]
name = "munra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygments>=2.19.2",
    "python-fasthtml>=0.12.35",
]
